test(activerecord): merge abstract/schema-definitions.test.ts into its .trails sibling
